Moved as_datetime_utc to tests/helpers
