+++ tests/helpers.py
@@ -3,8 +3,11 @@
 
 from __future__ import absolute_import
 import os
+from dateutil.parser import parse as dateutil_parse
 from dateutil.rrule import rrule, DAILY
 from packaging import version
+
+from django.utils.timezone import utc
 
 from organizations.models import Organization
 
@@ -65,3 +68,22 @@ def django_filters_pre_v2():
     """
     import django_filters
     return version.parse(django_filters.__version__) < version.parse('2.0.0')
+
+
+def as_datetime_utc(datetime_string):
+    """Returns `datetime` instance with UTC timezone
+
+    This helpler function centralizes converting  "datetime as a string" to a
+    datetime object in UTC.
+
+    We use dateutil.parser.parse because it is convenient. However, these tests
+    need to support mulitple versions of Open edX and with that, multiple versions
+    of the dateutil package. While the dateutil package does handle timezone
+    assignment, this approach is more portable.
+
+    We may want to iterate on this to create a 'datetime_matches' function.
+    However, then we have to consider type checking for the parameters, or
+    enforce one type as string and the other as datetime, or do conversions.
+    Basically, we might be making testing more complicated
+    """
+    return dateutil_parse(datetime_string).replace(tzinfo=utc)
